Changed size variable from unsigned int to unsigned long int

+++ src/config/syscheck-config.h
@@ -267,7 +267,7 @@ typedef struct registry_ignore_regex {
 
 typedef struct fim_file_data {
     // Checksum attributes
-    unsigned int size;
+    unsigned long int size;
 #ifdef WIN32
     cJSON * perm_json;
 #endif
